+++ src/loading.ts
@@ -15,14 +15,14 @@ export interface IShowLoadingConfig {
  */
 export const showLoading = (text = '请稍候', seconds = 10, config: IShowLoadingConfig = {}) => {
     const defaultConfig: IShowLoadingConfig = {
-        hasMask: true,
-        maskColor: 'transparent',
+        hasMask: false,
+        maskColor: '#00000070',
         onCancel: null,
         cancelInline: false,
         id: 'com_global_page_loading',
     };
     config = Object.assign({}, defaultConfig, config);
-    const { id } = config;
+    const { id, hasMask, maskColor, onCancel, cancelInline } = config;
     const timeoutKey = `_${id}_timeout`;
     if ((window as any)[timeoutKey]) {
         clearTimeout((window as any)[timeoutKey]);
@@ -41,8 +41,24 @@ export const showLoading = (text = '请稍候', seconds = 10, config: IShowLoadi
         style.innerHTML = `
         .${id} {
             position: fixed;
-            top: calc(50vh - 60px);
-            left: calc(50vw - 60px);
+            left: 0;
+            top: 0;
+            width: 100vw;
+            height: 100vh;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            pointer-events: none;
+            z-index: 10000;
+        }
+        .${id}.show-mask {
+            pointer-events: unset;
+            background: #00000070;
+        }
+        .${id} .mask-wrapper {
+            position: fixed;
+        }
+        .${id} .loading-wrapper {
             width: 120px;
             height: 120px;
             z-index: 8000;
@@ -50,32 +66,45 @@ export const showLoading = (text = '请稍候', seconds = 10, config: IShowLoadi
             border-radius: 8px;
             text-align: center;
             color: white;
-            padding-top: 20px;
+            display: flex;
+            justify-content: center;
+            flex-direction: column;
+            align-items: center;
+            pointer-events: initial;
+            line-height: 1.3;
         }
-        .${id} img {
+        .${id} .loading-wrapper img {
             width: 50px;
+            height: 50px;
             margin-bottom: 10px;
+        }
+        .${id} .loading-wrapper a {
+            color: #6ab2ff;
         }`;
         document.head.appendChild(style);
     }
     dom.innerHTML = `
-        ${config.hasMask ? `<div class="mask-wrapper" style="background-color: ${config.maskColor}"></div>` : ''}
         <div class="loading-wrapper">
-            <div class="loading-content">
-                <img src="https://img.alicdn.com/tfs/TB1bnUsQBLoK1RjSZFuXXXn0XXa-32-32.svg" alt="加载中">
-                <div>${text}${config.cancelInline ? ' ' : '</div>'}
-                ${config.onCancel ? '<a href="javascript:;" class="cancel">取消</a>' : ''}
-                ${config.cancelInline ? '</div>' : ''}
+            <img src="https://img.alicdn.com/tfs/TB1bnUsQBLoK1RjSZFuXXXn0XXa-32-32.svg" alt="加载中">
+            <div>
+                <span>${text}</span>
+                ${cancelInline ? '' : '<br />'}
+                ${onCancel ? '<a href="javascript:;" class="cancel">取消</a>' : ''}
             </div>
         </div>`;
-    if (config.onCancel) {
+    if (onCancel) {
         const btn = dom.querySelector('.cancel');
         btn && btn.addEventListener('click', () => {
             hideLoading();
-            config.onCancel();
+            onCancel();
         });
     }
-    dom.style.display = 'block';
+    dom.addEventListener('click', () => {
+        hideLoading();
+    });
+    dom.style.display = 'flex';
+    dom.style.background = hasMask ? maskColor : 'transparent';
+    dom.className = `${id} ${hasMask ? 'show-mask' : ''}`;
     if (seconds > 0) {
         (window as any)[timeoutKey] = setTimeout(() => {
             hideLoading();